mkdumprd: do not use lsinitrd to unpack initrd
In general, it is a good idea to reuse an existing utility...

However, dracut's lsinitrd does not preserve file modification
times. As a result files under /fadumproot are not hardlinked with
identical files from the normal initrd. And there's a _lot_ of
potential for de-duplication: On my test system, this change reduces
unpacked initrd size from 195 to 122 megabytes (almost 40%).

The behaviour of lsinitrd may change in the future, but that will
take time, so for now, let's re-implement initrd unpacking in
mkdumprd to make sure that file modification time is preserved.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>