Store makedumpfile-R.pl as data under /kdump in the initrd
This script is not supposed to be executed in the initrd
environment. It is conveniently saved to the target directory if
makedumpfile used the flattened format, so whoever wants to analyze
the dump later does not have to search for it.

The kdump initrd essentially treats it as data. Consequently, it
makes little sense to put the script with executable permissions
under /usr/bin. In fact, it is harmful, because dracut looks for all
executable files in the initrd and installs all dependencies. This
drags the perl binary into the initrd...

To fix the above-mentioned issues, store makedumpfile-R.pl under
/kdump and remove executable bits.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>