Remove undocumented config option KDUMPTOOL_OPTIONS
This undocumented option is not very useful. It is added at the end
of the kdumptool command line, so it can only be a subcommand
option, not a global option (such as "--debug"). It is used both
for the "delete_dumps" subcommand and the "save_dump" subcommand.
The only option these two subcommands have in common is "--root".
Although this option makes it possible to override the root
directory, it is hard to imagine a use case for it.

Let it RIP.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>