Rewrite save_kdump.sh in C++
The C++ version is much better at handling error conditions. It
also unifies config option parsing (e.g. kdumptool accepts "true"
to be an alias for "yes").

Last but not least, kdumptool is no longer needed in the kdump
initrd. In other words, the initrd no longer contains code that is
used only in the normal environment (e.g. "calibrate").

Signed-off-by: Petr Tesarik <ptesarik@suse.com>