Do not bother passing KDUMP options on the kernel command line
The rewrite of save_dump.sh revealed that the kernel command line
has never been used. The original script received all its
parameters from the initrd...

When the config file changes, the initrd must be re-generated
anyway, so the configuration therein is always up to date.

In short, do not mess with KDUMP options in the boot loader and
remove some unnecessary code. This is good (TM).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>