Create fadump initrd from the fadumpinit dracut module
The current mechanism works well if mkdumprd is used to build the
initrd. Unfortunately, there are other ways that dracut is invoked
automatically, e.g. from a post-install script when installing an
updated kernel package. If that happens, firmware-assisted dump
fails until the next mkdumprd invocation, because this new kernel
becomes the default boot option, but the initrd does not contain
fadump-specific bits.

So, let's do things differently.

The fadumpinit dracut module gets automatically included if
KDUMP_FADUMP is enabled in kdump configuration. Note that it can
still be explicitly omitted with a dracut command-line parameter.
The content of the fadump-specific initrd is then generated by the
install() function of this module, i.e. a second dracut instance is
executed recursively from a dracut module.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>