calibrate/maxrss.py: Processing of TASKSTATS data
This script should be used to post-process the output from trackrss
and find the maximum RSS over all processes. The algorithm is a bit
pessimistic, as it assumes that the maximum RSS for each process is
reached at startup and never decreased, but it's the only safe
option, because we don't get an overall maximum RSS figure from the
kernel, and periodic sampling may miss a short peak.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>