Run trackrss as PID 1
It turns out that other processes may already run when systemd is
executed. So, let's start trackrss even before systemd and leave a
tracking daemon around.

The standard fork-and-exec sequence looks somehow reversed here,
but that's necessary, because systemd wants to run with PID 1, and
we cannot really exit PID 1 anyway...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>