trackrss: Generate own timestamps
The accounting seems to work in general, but one-second resolution
is too coarse. Unfortunately, the kernel does not create any more
fine-grained timestamp of its own (SO_TIMESTAMP is not implemented
for netlink sockets), so best we can do is log an after-the-end
timestamp when the netlink message is received.

Doh.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>