trackrss: Read netlink in a separate thread
Now that the timestamping is done in userspace, it is important
that the netlink socket reads are not blocked by writing data to
the serial console.

I'm using a fixed-size ringbuffer with overflow detection.

The extra slot in the queue is used for data that is currently
being received. If the ringbuffer is full, this slot is discarded,
otherwise it is exchanged with the new head slot.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>