calibrate: Run a VM with two CPUs
Some kernel allocations are optimized out on UP systems, skewing
the memory requirements. In particular, the difference between
memory requirements of a UP system and a SMP system with 2 CPUs is
considerably bigger than between a SMP system with n CPUs and a
system with n+1 CPUs.

The simplest solution is to measure the base requirements on a
system with 2 CPUs.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>