calibrate: Open trackrss console early and clean it up afterwards
First, the device node is no longer needed after an open file
descriptor is available. Clean it up so it cannot cause any
incompatibility with systemd.

Second, if the console cannot be opened, it's probably better to
let the VM fail immediately. But it's too late for that from the
deamonized child.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>