calibrate: Abuse crashkernel= to protect the ELF core header
Unfortunately, memmap only works on ACPI systems, so it cannot be
used e.g. for IBM POWER...

There is a platform-independent option "reserve=", but that cannot
be used either, because it disallows reading the elfcorehdr itself
from that region.

It seems like a good idea to abuse the crashkernel= option. First,
it must be recognized on all platforms that support crash dumps and
second, the region is still regarded as normal RAM, just never
allocated to the running kernel, except when loading a crash
kernel, but we know that the test VM will not do that.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>