calibrate: Run poweroff AFTER saving kdump
The poweroff currently runs before even attempting the save. That's
before many user-space processes are even run. Consequently, the
user-space requirements are underestimated.

The issue is that we cannot save a real dump, not the least because
there is no crashed system. Creating a complete working kdump
environment would be a lot more work.

Moreover, there are no target devices. That's why the dummy config
specifies a network target; otherwise the VM would get stuck on
mounting target filesystems. However, if kdumptool is allowed to
run, it will wait for network that will not come up either...

Luckily, there is an (undocumented) configuration option to reduce
the network timeout, so the VM can terminate in some reasonably
short time even if poweroff becomes a POSTSCRIPT.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>