calibrate: Replace taskstat netlink with rss_stat event tracing
There is an issue with the RSS field reported by taskstats. The
value is a sum of file RSS, anonymous RSS and shared memory RSS,
but file RSS does not increase RAM consumption, because all rootfs
files are stored in RAM anyway.

Periodic sampling does not work either, because the goal is to
determine peak usage, but most processes are short-lived.

Luckily, there is a rss_stat ftrace event.

Note that trackrss can be simplified considerably. That's mostly
due to the fact that timing is no longer critical (as long as the
in-kernel ftrace ring buffer does not overflow).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>