calibrate: Improve mm_id matching
The mm_id is a randomized hash of the mm address. It is always zero
until there is enough entropy to generate the random seed. However,
that's always the case at boot.

For this test VM, it does not matter if the random seed is good. In
fact, we could just as well use the mm address directly, if the
kernel allowed that. Instead, let's just pretend there's enough
entropy...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>