calibrate: Run dracut with the uninstalled kdump module
Dracut does not offer many ways to use a module that is not
installed in the system-wide directory, but there is this --local
option, which is mainly used during the development of new dracut
features. Consequently, this mode it assumes that dracut is running
in a dracut git working directory, i.e. it requires that _all_
other files are found relative to the current directory. But we
want to add/replace only one module...

So, let's symlink everything from the system directory, except the
kdump module, and then symlink the kdump module from the temporary
directory where it had been installed.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>