calibrate: Generate calibrate.conf at build time and install it
The run-qemu.py script is quite heavy in terms of dependencies and
resource usage. It is not supposed to be run regularly on the
target system. Instead, run it only once at build time to find out
the memory consumption for the target OS.

The added dependency on dracut build system targets should avoid
attempts at running 'cmake --install' in the init/ subdirectory
before it is built. The BUILDSYSTEM_TARGETS property was added in
CMake 3.7, and this is why the minimum CMake version had to be
increased.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>