calibrate: Fix serial consoles in System z VMs
The s390x emulator cannot create multiple VT220 operator
consoles (/dev/sclpN). The secondary port must be a virtio serial
port. The situation is different from IBM POWER, where the output
files are are mapped to hvc0 and hvc1. On s390x, the kernel console
must stay on sclp0, because that's the only way to capture early
boot messages (with all the init mem information that is parsed by
kernel.py).

Signed-off-by: Petr Tesarik <ptesarik@suse.com>