calibrate: Use virtio-net with s390x
QEMU for s390x does not support the e1000e NIC model. It would be
possible to use virtio-net for all platforms, but an emulated
"real-world" NIC is more likely to match real-world requirements. Of
course, then we'd like to use QETH on s390x, but sadly, that's not
an implemented by QEMU. So, let's stay with virtio-net...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>