calibrate: Fix serial console for AArch64
QEMU does not implement multiple PL011 serial ports[1]. Luckily,
SUSE kernels build with CONFIG_8250_PCI=y, so PCI serial ports can
be used instead.

The PL011 serial port is created anyway, but it is redirected to a
null device, because the OVMF firmware prints a gazillion debug
messages to that port before the kernel is loaded.

[1] https://lists.nongnu.org/archive/html/qemu-discuss/2018-11/msg00001.html

Signed-off-by: Petr Tesarik <ptesarik@suse.com>