Do not store implicit host devices in the initrd
All devices that may be required for kdump are explicitly requested
by the kdump dracut module. In fact, some host devices may not even
exist in the kdump initrd. If dracut creates an initqueue check for
them, then initqueue may time out, eventually dropping to the
emergency shell.

Similarly, host devices will not be found in the calibration QEMU
guest.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>