calibrate: Account for kernel image size
The location of the kernel image may limit the maximum possible
size of the crash kernel area. With kASLR, the kernel may be placed
just in the middle of a memory block, leaving only a quarter of the
block available before and after the kernel image.

For now, let's assume that the kernel image is always placed in the
same RAM block, because this issue is most important for small
systems with only one resonably sized RAM block.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>