calibrate: Account for initrd size
The previous commit was a great improvement on small systems, but
it can get worse than that, because the boot loader allocates the
initrd separately, often in the middle of an otherwise available
memory area.

For example, in a small x86_64 VM with 1.5 GiB of RAM, the memory
at boot looks somewhat like this (kernel 5.16 loaded with GRUB
2.06):

00100000-5ffdbfff : System RAM
  00100000-22dfffff : Free RAM: 557 MiB
  22e00000-23c0209f : Kernel code (.text)
  23c020a0-23dfffff : small gap
  23e00000-24826fff : Kernel rodata (.rodata)
  24827000-249fffff : small gap
  24a00000-24d4fa7f : Kernel data (.data)
  24d4fa80-251a3fff : small gap
  251a4000-25215dd9 : Kernel init text (.init.text)
  25215dda-25217fff : small gap
  25218000-253fdfff : Kernel init data (.init.data)
  253fe000-25407fff : small gap
  25408000-259fffff : Kernel bss (.bss)
  25a00000-35356fff : Free RAM: 249 MiB
  35357000-369a2fff : RAMDISK
  369a3000-5ffdbfff : Free RAM: 662 MiB

The specifics of ramdisk allocation are out of scope, so let's
assume the worst possible fragmentation for the limit.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>