kdumptool calibrate: add more margin to reservation calculations
We're getting too many bugs where the autoconfigured crashkernel value provided
by kdumptool calibrate is not sufficient.

Currently 10% was added to the calculated value.

This patch changes this multiplicative constant (ADD_RESERVE_PCT) to to 20% and
introduces another additive constant ADD_RESERVE_KB, set to 8 MiB.

The goal is to make the defaults work for more setups at the expense of memory,
letting memory-frugal users to tune the default values down.