Fix network-related dracut options handling for fadump case
Commit bbe9b1281cd6 ("Do not add  network-related dracut options if
ip= is set explicitly") ensures kdump network-related dracut options
are set according to admin's intention. It relies on KDUMP_COMMANDLINE
& KDUMP_COMMANDLNE_APPEND config options for that. But these options
are currently not applicable for fadump case. Instead, rely on the
parameters passed to the running kernel, as capture kernel boot is
likely to use the same parameters as production kernel for fadump.

Fixes: bbe9b1281cd6 ("Do not add  network-related dracut options if ip= is set explicitly")
Signed-off-by: Hari Bathini <hbathini@linux.ibm.com>