kdumptool calibrate: modify fadump suggestions
Per jsc#IBM-1027, calibrate should output fadump values separately
from non-fadump values.

Drop all the code tweaking non-fadump values for use by fadump when fadump
is turned on.

Three new values shall be ouptput on platforms supporting fadump:
Fadump, MinFadump and MaxFadump.

If and only if these are present in the output of kdumptool calibrate,
the configuration tool (YaST) should present the user with an option tu switch on
fadump, and when selected, these values define the default, min and max values
for fadump memory reservation.

The values are:
	Fadump: 5% of total memory
	MinFadump: 64 MB
	MaxFadump:  50% of total memory