print the qemu messages and trackrss log during calibrate
When compiled with the calibrate option, debugging problems
inside the executed qemu is very difficult as no output is printed.
Fix this by running a tail -f on both TRACKRSS_LOG and MESSAGES_LOG.

Also disable parallel build in spec file by runing make direcly and not through
the %cmake_build macro, as this turns out the make output synchronization feature
and if qemu hangs, none of its output is printed during build.