kdumptool calibrate: add even more margin to reservation calculations
We're getting too many bugs where the autoconfigured crashkernel value provided
by kdumptool calibrate is not sufficient.

Currently 20% + 8 MB was added to the calculated value.

Make this 30% + 64 MB.

The calculated value omits too many small things (like makedumpfile) that
result in kdump not working by default far too often.

The valuable part of the calibration infrastructure is
the estimates needed to scale to huge machines.

The goal is to make the defaults work for more setups at the expense of memory,
letting memory-frugal users to tune the default values down.

Commit e7151803 apparently did not go far enough.