Simplify kdump-save and initrd generation
This is the first of a series of changes aiming to substantially simplify
the kdump code base.

This is complete rewrite of kdump-save in bash achieving simplification by
deprecation of obsolete or complicated fatures and using existing dracut
functionality and external programs instead of our own re-implementaion in
kdump-save / kdumptool.

Following commits will clean up and remove most of the kdumptool code, most of
which is now unused.

Features deprecated:

- split dumps (saving to more than one targets at once)

  Originally this feature allowed to speed things up by multithreading and
  combining I/O bandwidth of more block devices.  The former is now obsoleted
  by multithreading support in makedumpfile.  Combining I/O badwidth seems too
  much of a corner case to keep things complicated.  Users with special needs
  like this can always write their own script to do exactly what they need and
  use it in KDUMP_TRANSFER

- KDUMPTOOL_FLAGS option removed

  Most of the flags dealt with the split dumps.
  XENALLDOMAINS is now default and can be disabled by including -X in
  MAKEDUMPFILE_OPTIONS

- notification e-mails

  Sending e-mail from the memory-restricted kdump environment is not ideal.
  The plan is to implement a simple notification service to be run after boot
  in the normal OS, checking for locally stored dumps. Network-stored dumps
  can be detected by the hosts that keep them.

- copying of the kernel image (KDUMP_COPY_KERNEL)

  The kernel image itself, as shipped in SUSE distributions, is not enough for
  debugging because it needs to be accompanied by the debug info. Determining
  which kernel to copy and mounting the respective file system is not easy and
  with the very limited usefulness is not worth the trouble.

These changes allow a huge simplification of everything pertaining to mounting
file systems. The only mounts needed are for the file, nfs and cifs targets.
This can be done by dracut --mount option and allowed removal of large portions
of the kdump dracut module code.

FTP and SFTP are now handled by lftp, added to the spec file as Recommends:

As with the other protocols, both SSH and SFTP now support passwords provided
in the URL.

KDUMP_SSH_HOST_KEY, previously broken, is now fixed but requires the key type to be
specified together with the key itself.

New option, KDUMP_DUMPFORMAT=raw, will save an unmodified /proc/vmcore
without any processing with makedumpfile.

The output directory name is now YYYY-MM-DD-HH-MM, i.e. the separator between
HH and MM changed from a colon to a dash (colons have issues on CIFS shares and
are uncomfortable to work with in  bash)

Unified the default KDUMP_SAVEDIR across the default config file, code and man
page to /var/crash, as the historic move from /var/log/dump to /var/crash was
somewhat unfinished.

ping is now used to detect network is with a timeout of KDUMP_NET_TIMEOUT seconds.
Set KDUMP_NET_TIMEOUT to 0 if ping does not work in your environment.

All the boolean / yesno options in /etc/sysconfig/kdump have been unified to
boolean, with canonical values "true" / "false". "yes", "no", "0" and "1" are still
accepted as valid values and converted to "true" / "false".