build initrd in /var/lib/kdump, use default kernel
Put generated kdump initrd in /var/lib/kdump/initrd. When the initrd is
generated, put a symlink to the associated kernel image in
/var/lib/kdump/kernel

This way load.sh can use these fixed files. All the logic determining which
kernel to use is moved to mkdumprd.

Allow absolute path or kernel version string to be specified in
KDUMP_KERNELVER. If it isn't, just use whichever default image is found at
/boot/{vmlinuz,image,Image,vmlinux}

KDUMP_KERNELVER can be overridden with -k or -K option to mkdumprd.

This makes kdumptool find_kernel obsolete.

Drop the removal of --args-linux, it's completely obsolete.