fadump: truncate existing initrd
When FADUMP is enabled, /var/kdump/initrd is an empty file, it's only purpose
is to remember the last time mkdumprd was run and the initrds in /boot
regenerated.

When FADUMP was previously turned off, /var/kdump/initrd will contain  a real
initrd created for the non-FADUMP case. Now with FADUMP on, mkinitrd will just
touch this file and keep its useless contents, which is confusing and a waste of
space. Replace touch with > to make the file empty.