fadump: don't use lsinitrd to unpack initrd, again
e014f28209589b244561bfeeeb178ad3897824b5 was absolutely right:
We need to preserve modification times for the final hardlink,
performed by dracut, to hardlink most of the fadump initrd with the
standard initdd.

I missed that when converting back to lsinitrd in
eacf45bcb869472fd9150a6cd3e6f833715229fd.

However, with --no-compress and --no-early-microcode passed to dracut
when generating the fadump initrd, a simple cpio call is enough.
The decompression code is no longer needed.