prevent dracut emergency shell
Only allow dracut to run the emergency shell when kdump-save may run
a shell on its own (based on the setting of KDUMP_CONTINUE_ON_ERROR,
KDUMP_IMMEDIATE_REBOOT and KDUMP_FADUMP_SHELL).

Otherwise make dracut reboot instead to prevent kdump hanging
in an emergency shell when an interactive shell is not expected.