man: bind mounts are not supported
The path leading to KDUMP_SAVEDIR should not contain a bind mount.
Verified that this did not work either before the v1.9 rewrite.

I tried to fix this by having findmnt output FSROOT
and passing this as X-mount.subdir mount option.
But this breaks btrfs submvolmes, because
findmnt has a really strange semantics and mixes
btrfs subvolumes with bind mounts in the output of FSROOT.

So until there's need for working around this, just document
that bind mounts don't work.