Always use nr_cpus, not maxcpus
PPC was the only supported platform to still use maxpus to limit
number of CPUs in the kdump environment. This caused issues with
multiqueue block drivers, likely since kernel commit d802057c7c55.

A patch fixing nr_cpus on PPC is on the way to the upstream kernel:
https://lore.kernel.org/all/170800202447.601034.7290612623478478380.b4-ty@ellerman.id.au/#t

Use nr_cpus on all architectures.

This patch has to be present in the kernel for nr_cpus to work on PPC:
https://git.kernel.org/pub/scm/linux/kernel/git/powerpc/linux.git/commit/?id=0875f1ceba974042069f04946aa8f1d4d1e688da

Fixes: bsc#1218180