init-fadump: remove root= from cmdline for systemd
The systemd fstab generator will mount root if found on the
kernel command line.

In kdump we prevent this by setting root=kdump rootflags=bind. However, in
fadump we can't do that as we boot with the standard command line used by the
production system.

The init-fadump wrapper is used as /init in the initrd.
It already mounts /proc to determine if we're booting fadump or standard boot.

Extend it to also get /proc/cmdline and replace root=... with root=kdump rootflags=bind
This adjusted cmdline is the passed to systemd using the SYSTEMD_PROC_CMDLINE
environment variable.

References: bsc#1219958