mask the dracut-rootfs-generator
In kdump/fadump we don't want to mount root.

It wastes memory and can conflict with mounting the target directory if they
are on the same filesystem (the root is mounted r/o and depending on which is
mounted first the target may end up mounted r/o as well).

For kdump, this is ensured by passing root=kdump rootflags=bind on the commandline.

For fadump we have no control over the command line. Commit
4c3e02548f6b12d00d7088e61c67c8258d405c01 ensures that the systemd rootfs
generator does not mount root by passing an alternative commandline to systemd
using SYSTEMD_PROC_CMDLINE.

But dracut has another systemd generator: dracut-rootfs-generator
This one creates initqueue jobs to wait for the root device and
a fsck job to check the root filesystem. For root=kdump it does nothing.
Mask it to prevent it being run even for fadump.

References: bsc#1219958