use the same persistent device names as dracut
Currently kdump gets the persistent device names from
udevadm info and gets the by-path or mapper symlinks.

This has a couple of drawbacks:

- The symlinks obtained from udevadm are in random order,
  making the behaviour unstable - each time the initrd
  is regenerated a different symlinks can be used; in case
  there is a problem, the bug is hard to reproduce

- mapper/ and by-path/ symlinks are used for the kdump fstab
  however dracut prefers mapper/, by-uuid/, by-label/ and other
  kinds before by-path/
  The result is that both schemes are used in the same initrd:
  by-path in the fstab and (usually) by-uuid in the dracut initqueue.
  If there is a problem with either one of the schemes kdump fails.

Use the persistent device code from dracut. This will make the fstab
and the initqueue use the same scheme.

Also, by-path is last in dracut's order for a reason - it is known
to fail in certain multipath setups. Changind to the dracut order will
fix kdump on these setups.

Fixes: b61504b74c3fbf782e4e1c023759a1ae75b0bce7