prevent busy loop when DNS fails
The kdump-save script uses ping in a loop to wait up to KDUMP_NET_TIMEOUT
seconds for the remote host to be accessible. In some cases, like a DNS
problem, the ping command fails immediately and the loop busy-spins really
fast, cluttering the output.

Sleep for a second when ping fails.