fix bonding options
kdump determines bonding options by parsing the output of "ip -d link show"
and leaving out a few of the reported options that are read-only.

Unfortunately, not all options are applicable for every bonding mode.

Depending on the network module handling the network configuration in the
initrd, the superfluous options are either ignored (network-legacy) or cause an
error (networkmanager).

Instead of adding all the options, make a list of options applicable for the
current bonding mode.

Read the option values from sysfs.

Don't include options that are set to their default value.

Some options need special handling:
"primary" needs a valid slave name. Make sure the name is the one used in the correct
slave name that will be used in the initrd, not the current slave name.

arp_ip_target separators need to be changed into semicolons.

ns_ip6_target is not currently supported by dracut, but replace the separators
as well just to avoid potentially breaking the command line syntax

References: bsc#1235933