Set default KDUMP_CPUS to 32
Commits 75834c3f692f26139e12a08e162569e6e211ed11 and
60525b4afa680650c9c85b4dc26372b34481d908 changed the default KDUMP_CPUS from
1 to 0 (meaning all CPUs). This was requested by IBM in jsc#PED-9894, bsc#1237754.

This change wa short-sighted: on machines with very large numbers of CPUs this
default does not scale. The crashkernel memory requrements grow beyond
reasonable. bsc#1240769 shows a system with 640 CPUs which does not work unless
1.8 GB memory is reserved.

Change the default to 32. This should require at most a couple tens of MBs extra
reserved memory compared to the original default of 1.

30-40 CPUs should also roughly be enough to fully load a PCIe device such as a
disk controller or a NIC used to save the dump.  A similar result can be seen
in the graph provided in the of commit 75834c3f692f26139e12a08e162569e6e211ed11
- the performance did not improve with more than ~30 threads.

Fixes: 75834c3f692f26139e12a08e162569e6e211ed11
Fixes: 60525b4afa680650c9c85b4dc26372b34481d908