check/update kernel command line from kdump-commandline.service
Introduce a new systemd service: kdump-commandline.service The purpose is to
check and possibly update the kdump-related command line options (crashkernel
and fadump) on the kernel commandline.

The reason for not doing this from the existing kdump.service is to be possible
to fail without necessarily failing kdump itself, serving as a warning
prompting the user to check or update the commandline options.

Introduce a KDUMP_UPDATE_BOOTLOADER option and
new options -D and -U to kdump commandline, conditionally doing the
same as -d and -u based on the value of KDUMP_UPDATE_BOOTLOADER.

Call kdumptool commandline -D during rpm uninstall.

Remove the old fadump bootloader config code from init/load.sh
as this is now handled by the new service.

References: jsc#PED-12454