Add caution for KDUMP_CPUS configuration
Setting KDUMP_CPUS to a very high value - for example, 900 or more can
cause the kdump/fadump kernel to fail during boot due to limited
system resources in the dump capture environment.

Add a caution note in the configuration file and the man page to warn
users about the potential consequences of setting KDUMP_CPUS to such
a high value.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>