fix and improve calibrate build
The calibrate build (using the --with calibrate option) uses qemu to run a
virtual machine running an intstrumented kdump environment to measure memory
consumption.

Two runs are done - one emulating a disk dump target and another one emulating a
network dump target.

The emulation was not perfect: it was expected that the dump itself would fail,
because the vmcore file is a fake one and makedumpfile won't process it. No
actual disk was supplied, no actual network connection was attempted.

Network setup relied on dhclient and the legacy dracut module, while network manager would
be used in a moder system.

Most importantly nothing checked how far the dumping emulation got before
finally failing; lately e.g. the network dump would fail as early as not
loading the network card driver, because proper udev was not present.

The result was really inaccurate measurements.

Fix the above issues:

- emulate the whole dumping process, using /proc/kcore instead of /proc/vmcore
- dump to a real emulated disk and a real network sftp server
  (start a ssh server on the build host)
- check if the dump was successful (the disk image or host system actually contain
  a finished dump with a proper vmcore and README.txt); fail if not to prevent silent failures
- revised spec file dependencies (NM, udev, systemd, ssh, e2fsprogs)
- fix s390x oldmem parameters in the parmarea
- separate the standard console log and the trackrss output to improve readability

The process newly requires building as root (to be able to mount the disk image
to verify the dump result and to start the ssh server). That's why needsrootforbuild is
added to the spec file and "BuildFlags: allowrootforbuild:1" is now needed in the project config
besides the definition of with_calibrate