Exclude specific panic sysctls from capture kernel initrd
The following sysctl configurations are retained in the capture kernel
initrd, which can cause dump capture failures unnecessarily:

- kernel.panic_on_warn
- kernel.softlockup_panic
- kernel.panic_on_rcu_stall

For example, if kernel.panic_on_warn is set in the capture kernel and a
warning is triggered, the capture kernel will panic, causing the dump
capture to fail.

The same applies to the other two sysctl configurations. The capture
kernel should avoid panicking on soft-lockups and RCU stalls, as the
system might still recover.

Fix this by excluding the above-mentioned sysctl configurations from
being retained in the capture kernel initrd.

Reported-by: Misbah Anjum N <misanjum@linux.ibm.com>
Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>